LPS-54017 Match column type to WorkflowDefinitionLink table

+++ portlets/kaleo-forms-portlet/docroot/WEB-INF/src/com/liferay/portal/workflow/kaleo/forms/portlet/KaleoFormsAdminPortlet.java
@@ -376,7 +376,7 @@ public class KaleoFormsAdminPortlet extends BaseKaleoFormsPorlet {
 		long ddmTemplateId = ParamUtil.getLong(actionRequest, "ddmTemplateId");
 		String workflowDefinitionName = ParamUtil.getString(
 			actionRequest, "workflowDefinitionName");
-		long workflowDefinitionVersion = ParamUtil.getLong(
+		int workflowDefinitionVersion = ParamUtil.getInteger(
 			actionRequest, "workflowDefinitionVersion");
 		String taskFormPairsData = ParamUtil.getString(
 			actionRequest, "taskFormPairsData");
